Define my modifyTodo method in every place i'm using it
